refactor(android): move JokerDisplayResolver out of components package for sonar coverage
